Add Gradle support to Pix.
