+++ src/main/java/Pix/exception/PixNumberFormatException.java
@@ -1,0 +1,11 @@
+package Pix.exception;
+
+/**
+ * Triggers when there is a number format Pix.exception.
+ */
+public class PixNumberFormatException extends PixException {
+    @Override
+    public String getMessage() {
+        return "Put the number as an integer!";
+    }
+}
