--- src/main/java/exception/PixIOException.java
@@ -1,11 +1,0 @@
-package exception;
-
-/**
- * Triggers when there is an I/O error in the program.
- */
-public class PixIOException extends PixException {
-    @Override
-    public String getMessage() {
-        return "There is an I/O error!";
-    }
-}
