--- src/main/java/exception/PixNoSuchTaskException.java
@@ -1,11 +1,0 @@
-package exception;
-
-/**
- * Triggers when there is no such task in the task list.
- */
-public class PixNoSuchTaskException extends PixException {
-    @Override
-    public String getMessage() {
-        return "There is no such task in the task list! Are you sure you can type?";
-    }
-}
