--- src/main/java/exception/PixNumberFormatException.java
@@ -1,11 +1,0 @@
-package exception;
-
-/**
- * Triggers when there is a number format exception.
- */
-public class PixNumberFormatException extends PixException {
-    @Override
-    public String getMessage() {
-        return "Put the number as an integer!";
-    }
-}
